Adding first print iterations
After first trying to print without supports of a brim or raft. After
a few failures and making sure the bed was level, made some revisions.

I checked the size and made the inner radius larger and added supports
and a raft when using flashforge. This printed well and apart from a
small layer-tearing when removing from the supports looked good. When
I increased the radius the arc length also increased enough to make it
overhang :/

I roughed the headphones ( GID6B) and the new reinforcement and glued
with Loctite epoxy. After initial setting, I tried it on. I didn't let
epoxy set fully and it peeled up at the previous break. Even after
this, the the reinforcement applies too much pressure on my
ears.

Next revision will be less rigid with less infill and less of an arc.